Implemented price range filter and selected filters above products

+++ server/auctionapp/src/main/java/com/atlantbh/auctionapp/projections/BidProj.java
@@ -2,12 +2,11 @@ package com.atlantbh.auctionapp.projections;
 
 import com.atlantbh.auctionapp.model.UserEntity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface BidProj {
     long getId();
-    BigDecimal getPrice();
+    double getPrice();
     LocalDateTime getBidDate();
     UserEntity getUser();
 }
